+++ packages/react/package.json
@@ -35,13 +35,11 @@
     "publish:dist": "rm -rf dist/.cache; cp ./{LICENSE,README.md,package.json} ./dist; cd ./dist; npm publish"
   },
   "peerDependencies": {
-    "@emotion/css": "^11.7.1",
     "@unovis/ts": "1.0.3",
     "react": ">=16.8.0 || ^17 || ^18",
     "react-dom": ">=16.8.0 || ^17 || ^18"
   },
   "devDependencies": {
-    "@emotion/css": "^11.7.1",
     "@rollup/plugin-node-resolve": "^13.0.4",
     "@types/react": ">=16.8.0",
     "@types/node": "^16.11.17",
@@ -51,6 +49,8 @@
     "rimraf": "^3.0.2",
     "rollup": "^2.61.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-rename-node-modules": "^1.3.1",
     "rollup-plugin-typescript2": "^0.31.1",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
     "tslib": "^2.3.1",
