+++ packages/react/rollup.config.js
@@ -3,6 +3,8 @@ import resolve from '@rollup/plugin-node-resolve'
 import peerDepsExternal from 'rollup-plugin-peer-deps-external'
 import typescript from 'rollup-plugin-typescript2'
 import transformPaths from '@zerollup/ts-transform-paths'
+import postcss from 'rollup-plugin-postcss'
+import renameNodeModules from 'rollup-plugin-rename-node-modules'
 import pkg from './package.json'
 
 // Array of extensions to be handled by babel
@@ -30,11 +32,15 @@ export default {
     resolve({
       extensions,
     }),
+    postcss({
+      modules: true,
+    }),
     typescript({
       typescript: require('typescript'),
       tsconfig: './tsconfig.lib.json',
       transformers: [(service) => transformPaths(service.getProgram())],
     }),
+    renameNodeModules(),
   ],
   external: regexesOfPackages,
 }
